Closes #30
Now only one enemy can be killed per one projectile.

In addition, note that the projectile closest to the enemy will be the one
that kills the enemy, since the one that was shot earlier has a lower index
and thus will be checked first in the loop. How convenient!